BC-70: Moved LaboratorySupervisorField from Organisation to Laboratory
